Validate task parameters in Ingestor using task schemas
